Change logo back to backround, and scale so it keeps ratio.
Inserting it as an <svg> element in the DOM made it part of the content
flow, this way it is a background image and behaves more nicely if you
want to print the page or do other strange things.
Also the scaling is now done in a way that it always keeps it's
intended proportions.